+++ internal/skysql/provisioning/update_service_node_number.go
@@ -1,5 +1,6 @@
 package provisioning
 
 type UpdateServiceNodesNumberRequest struct {
-	Nodes int64 `json:"nodes,omitempty"`
+	Nodes         int64  `json:"nodes,omitempty"`
+	MaxscaleNodes *int64 `json:"maxscale_nodes,omitempty"`
 }